added phone number and fixed database lock
