+++ .gitignore
@@ -62,6 +62,7 @@ db.sqlite3
 db.sqlite3-journal
 
 # Flask stuff:
+*.db
 instance/
 .webassets-cache
