+++ Pipfile
@@ -14,7 +14,8 @@ MarkupSafe = "==3.0.2"
 Werkzeug = "==3.1.3"
 stripe = "==12.2.0"
 python-dotenv = "==1.0.1"
-pytz = "*"
+requests = "==2.32.4"
+pytz = "==2025.2"
 
 [dev-packages]
