+++ app/templates/register.html
@@ -132,6 +132,10 @@
                 <label for="confirm" class="form-label">Підтвердження пароля</label>
                 <input type="password" class="form-control" id="confirm" name="confirm" placeholder="Повторіть пароль" required>
             </div>
+            <div class="mb-3"> {# New phone number input field #}
+                <label for="phone_number" class="form-label">Номер телефону</label>
+                <input type="tel" class="form-control" id="phone_number" name="phone_number" placeholder="Введіть номер телефону (+380XXYYYYYYY)" required>
+            </div>
             <button type="submit" class="btn btn-primary w-100">Зареєструватися</button>
         </form>
